requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-CORS==4.0.2
Werkzeug==3.0.6
SQLAlchemy==2.0.23
python-dotenv~=1.0.1
pandas~=2.2.2
tqdm~=4.66.5
psutil~=5.9.0
psycopg2-binary~=2.9.10
isodate~=0.6.1
numpy~=1.26.4
pyspellchecker~=0.8.2
flask-caching~=2.0.1